+++ app/App/Account/Account/Requests/TransactionRequest/TxMain/index.js
@@ -6,7 +6,7 @@ import link from '../../../../../../../resources/link'
 import svg from '../../../../../../../resources/svg'
 import { Cluster, ClusterRow, ClusterValue } from '../../../../../../../resources/Components/Cluster'
 import { getAddress } from '../../../../../../../resources/domain/transaction'
-import { displayValueData } from '../../../../../../../resources/domain/transaction/displayValue'
+import { displayValueData } from '../../../../../../../resources/utils/displayValue'
 
 class TxSending extends React.Component {
   constructor (...args) {
