+++ test/resources/utils/displayValue.test.js
@@ -1,5 +1,5 @@
 import { BigNumber } from 'bignumber.js'
-import { displayValueData } from '../../../../resources/domain/transaction/displayValue'
+import { displayValueData } from '../../../../resources/utils/displayValue'
 
 describe('wei', () => {
   it('should return a wei value', () => {
